[FIX/#22] Selfie 프래그먼트로 변환
